Rename all the auth to login
