+++ api/src/handlers/login.js
@@ -23,11 +23,11 @@ const replyWithInvalidCreds = async reply => {
   reply(boom.unauthorized('invalid username or password', 'jwt'));
 };
 
-const auth = {
-  login: {}
+const login = {
+  post: {}
 };
 
-auth.login.config = {
+login.post.config = {
   validate: {
     payload: {
       email: joi.string().email().required(),
@@ -36,7 +36,7 @@ auth.login.config = {
   }
 };
 
-auth.login.handler = async (request, reply) => {
+login.post.handler = async (request, reply) => {
   const {email, password} = request.payload;
   const {jwtAudience, jwtPrivateKey} = request.config.get('auth');
 
@@ -59,4 +59,4 @@ auth.login.handler = async (request, reply) => {
   }
 };
 
-module.exports = auth;
+module.exports = login;
